Split serverless into independent directory and github-actions runner
