+++ .github/workflows/ci.yml
@@ -51,11 +51,26 @@ jobs:
       - name: Check out code
         uses: actions/checkout@v2
 
+      - name: Test
+        run: make test-e2e
+
+  integration-tests-serverless:
+    name: Test serverless integration e2e suite
+    runs-on: ubuntu-latest
+    steps:
+      - name: Set up Go 1.17
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.17
+
+      - name: Check out code
+        uses: actions/checkout@v2
+
       - name: Add buildpack commands for serverless testing
         uses: buildpacks/github-actions/setup-pack@v4.5.0
 
       - name: Test
-        run: make test-e2e
+        run: make test-e2e-serverless
 
   build:
     name: Build
@@ -130,7 +145,7 @@ jobs:
           curl -fSL -o jsonnet.tar.gz https://github.com/google/jsonnet/releases/download/v0.17.0/jsonnet-bin-v0.17.0-linux.tar.gz
           tar -xvf jsonnet.tar.gz -C /usr/local/bin/
           chmod a+x /usr/local/bin/jsonnet
-          
+
           go install github.com/jsonnet-bundler/jsonnet-bundler/cmd/jb@v0.4.0
 
           curl -fSL -o /usr/local/bin/tk https://github.com/grafana/tanka/releases/download/v0.19.0/tk-linux-amd64
